+++ src/pages/calendar/booking-calendar/index.tsx
@@ -11,6 +11,8 @@ import dateRangeIterator from "../../../utils/dateRangeIterator";
 import { apartmentById } from "../../../types/apiData/apartment";
 import ApartmentSingleSearch from "../../../components/inputs/search/apartment-single-search";
 import Select from "../../../components/inputs/select";
+import useGetApartmentCalendar from "../../../services-hooks/apartmentCalendar";
+import useGetLocationGroupings from "../../../services-hooks/apartment/useGetLocationGroupings";
 
 const breadCrumb = [
   {
@@ -31,39 +33,25 @@ const nextMonthString = new Date(today?.getFullYear(), today?.getMonth() + 6, 0)
   ?.toISOString()
   ?.slice(0, 10);
 
-const dummyBookedDates = [
-  "2025-02-01",
-  "2025-02-06",
-  "2025-03-12",
-  "2025-04-13",
-  "2025-05-21",
-] as any;
-const dummyBlockedDate = [
-  "2025-01-21",
-  "2025-03-23",
-  "2025-04-23",
-  "2025-02-07",
-  "2025-05-14",
-  "2025-06-22",
-  "2025-02-25",
-] as any;
+// const dummyBookedDates = [
+//   "2025-02-01",
+//   "2025-02-06",
+//   "2025-03-12",
+//   "2025-04-13",
+//   "2025-05-21",
+// ] as any;
+// const dummyBlockedDate = [
+//   "2025-01-21",
+//   "2025-03-23",
+//   "2025-04-23",
+//   "2025-02-07",
+//   "2025-05-14",
+//   "2025-06-22",
+//   "2025-02-25",
+// ] as any;
 
 export default function ApartmentCalendarPage() {
   const dispatch = useAppDispatch();
-  // update page props on component mount
-  useLayoutEffect(() => {
-    dispatch(
-      updatePageProperties({
-        breadCrumb,
-        pageTitle: "Booking Calendar",
-        pageDescription: "Booking calendar",
-        isLoading: false,
-        failedToLoad: false,
-        setFailedToLoad: false,
-        retryRequest: false,
-      })
-    );
-  }, []);
 
   // component states
   const [apartment, setApartment] = useState<apartmentById>({} as any);
@@ -78,6 +66,32 @@ export default function ApartmentCalendarPage() {
 
   const [calendarVewData, setCalendarViewData] = useState<Date[]>([]);
 
+  // calendar data fetching based on filtered dates
+  const { data, isLoading, isFailed, retryFunction, setIsFailed } =
+    useGetApartmentCalendar({
+      id: String(apartment?.id || ""),
+      start_date: filterDates?.start_date,
+      end_date: filterDates?.end_date,
+    });
+  const { data: locationGroups } = useGetLocationGroupings({
+    page: 1,
+    limit: 1000,
+  });
+
+  // update page props on component mount
+  useLayoutEffect(() => {
+    dispatch(
+      updatePageProperties({
+        breadCrumb,
+        pageTitle: "Booking Calendar",
+        pageDescription: "Booking calendar",
+        isLoading: isLoading,
+        failedToLoad: isFailed,
+        setFailedToLoad: setIsFailed,
+        retryRequest: retryFunction,
+      })
+    );
+  }, [isLoading, isFailed, setIsFailed, retryFunction]);
   //spread date range into objects for the calendar view to render
   const extractCalendarViewData = useCallback(() => {
     if (filterDates?.start_date && filterDates?.end_date) {
@@ -133,7 +147,15 @@ export default function ApartmentCalendarPage() {
               <option value={""} disabled>
                 Filter by building
               </option>
-              <option value={"all"}>All</option>
+              <option value={""}>All</option>
+              {locationGroups.map((item) => (
+                <option
+                  key={String(item?.id || "")}
+                  value={String(item?.id || "")}
+                >
+                  {item?.name}
+                </option>
+              ))}
             </Select>
           </div>
           <Filter actionHandler={handleSalesFiltering} />
@@ -150,8 +172,8 @@ export default function ApartmentCalendarPage() {
                     <CalendarView
                       key={item?.toString() || index}
                       date={new Date(item || "")}
-                      booked={dummyBookedDates} //TODO: replace with real api data
-                      blocked={dummyBlockedDate} //TODO: replace with real api data
+                      booked={data?.booked_dates || []}
+                      blocked={data?.blocked_dates || []}
                       onDateClick={handleDateClick}
                       allowNaviagtor={false}
                     />