+++ src/services-hooks/bookings/useGetDepartingReservation.tsx
@@ -5,6 +5,8 @@ import {
 } from "../../stores/apiData/bookings/departing-reservation";
 import { useAppDispatch, useAppSelector } from "../../stores/hooks";
 import useAxios from "../../useHooks/useAxios";
+import ApiQueryParamsExtractor from "../../utils/api-query-params-extractor";
+import { pagination } from "../../types/pagination";
 
 //axios instace interceptor for access token integration and refresh tokens
 export default function useGetDepartingReservation({
@@ -30,41 +32,32 @@ export default function useGetDepartingReservation({
   const [isLoading, setIsLoading] = useState(false);
   const [isFailed, setIsFailed] = useState(false);
 
-  const [pagination, setPagination] = useState<{
-    current_page: number;
-    last_page: number;
-    per_page: number;
-    total: number;
-    from: number;
-    to: number;
-  }>({} as any);
+  const [pagination, setPagination] = useState<pagination>({} as any);
 
   const getDepartingReservation = useCallback(async () => {
     setIsLoading(true);
     setIsFailed(false);
     try {
+      const { queryString, remakeRequest } = ApiQueryParamsExtractor({
+        dataset: {
+          page: search ? 1 : page,
+          start_date: start_date,
+          end_date: end_date,
+          sort: sort,
+          search: search,
+        },
+      });
       //check store if this requested data has been saved previously and retirve it
       //if not, make a new request and save into store
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (
-        foundPage &&
-        !(start_date && end_date) &&
-        !(sort === "asc") &&
-        !search
-      ) {
+      if (foundPage && !remakeRequest) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateDepartingReservationList({ data: foundPage?.data }));
       } else {
         const response = await axios.get(
-          start_date && end_date
-            ? `/admin/booking/departing-reservation?sort=${sort}&limit=20&search=${
-                search || ""
-              }&page=${page}&start_date=${start_date}&end_date=${end_date}`
-            : `/admin/booking/departing-reservation?sort=${sort}&limit=20&search=${
-                search || ""
-              }&page=${page}`
+          `/admin/booking/departing-reservation??${queryString}`
         );
         const { bookings } = response?.data?.data;
         const { data, current_page, last_page, per_page, total, from, to } =
@@ -79,7 +72,7 @@ export default function useGetDepartingReservation({
           length: data?.length,
         };
         dispatch(updateDepartingReservationList({ data }));
-        if (!search) {
+        if (!remakeRequest) {
           dispatch(
             addToPaginationHistory({
               pagination_data: paginationDataset,