+++ src/services-hooks/reports/useGetAllOwnersReport.tsx
@@ -59,7 +59,7 @@ export default function useGetAllOwnersReport({
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (foundPage && !remakeRequest && !(sort === "asc")) {
+      if (foundPage && !remakeRequest) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateOwnersReportReport({ data: foundPage?.data }));
       } else {
@@ -77,7 +77,7 @@ export default function useGetAllOwnersReport({
           length: data?.length,
         };
         dispatch(updateOwnersReportReport({ data }));
-        if (!search) {
+        if (!remakeRequest) {
           dispatch(
             addToPaginationHistory({
               pagination_data: paginationDataset,