+++ src/services-hooks/useGetAllAdmins.tsx
@@ -6,6 +6,7 @@ import {
 } from "../stores/apiData/admins-list";
 import useAxios from "../useHooks/useAxios";
 import { pagination } from "../types/pagination";
+import ApiQueryParamsExtractor from "../utils/api-query-params-extractor";
 
 //axios instace interceptor for access token integration and refresh tokens
 export default function useGetAllAdmins({
@@ -34,31 +35,28 @@ export default function useGetAllAdmins({
   const [pagination, setPagination] = useState<pagination>({} as any);
 
   const getAllAdmins = useCallback(async () => {
+    setIsLoading(true);
+    setIsFailed(false);
     try {
-      setIsLoading(true);
+      const { queryString, remakeRequest } = ApiQueryParamsExtractor({
+        dataset: {
+          page: search ? 1 : page,
+          start_date: start_date,
+          end_date: end_date,
+          sort: sort,
+          search: search,
+        },
+      });
       //check store if this requested data has been saved previously and retirve it
       //if not, make a new request and save into store
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (
-        foundPage &&
-        !(start_date && end_date) &&
-        !(sort === "asc") &&
-        !search
-      ) {
+      if (foundPage && !remakeRequest) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateAdminsList({ data: foundPage?.data }));
       } else {
-        const response = await axios.get(
-          start_date && end_date
-            ? `/admin/admins?sort=${sort}&limit=20&search=${
-                search || ""
-              }&page=${page}&start_date=${start_date}&end_date=${end_date}`
-            : `/admin/admins?sort=${sort}&limit=20&search=${
-                search || ""
-              }&page=${page}`
-        );
+        const response = await axios.get(`/admin/admins?${queryString}`);
         const { admins } = response?.data?.data;
         const { data, current_page, last_page, per_page, total, from, to } =
           admins;
@@ -72,12 +70,14 @@ export default function useGetAllAdmins({
           length: data?.length,
         };
         dispatch(updateAdminsList({ data }));
-        dispatch(
-          addToPaginationHistory({
-            pagination_data: paginationDataset,
-            data: data,
-          })
-        );
+        if (!remakeRequest) {
+          dispatch(
+            addToPaginationHistory({
+              pagination_data: paginationDataset,
+              data: data,
+            })
+          );
+        }
         setPagination(paginationDataset);
       }
       setIsLoading(false);