+++ src/services-hooks/useGetAllBookingsLists.tsx
@@ -65,7 +65,7 @@ export default function useGetAllBookingsLists({
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (foundPage && !remakeRequest && !(sort === "asc")) {
+      if (foundPage && !remakeRequest) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateBookingsList({ data: foundPage?.data }));
       } else {
@@ -83,7 +83,7 @@ export default function useGetAllBookingsLists({
           length: data?.length,
         };
         dispatch(updateBookingsList({ data }));
-        if (!search) {
+        if (!remakeRequest) {
           dispatch(
             addToPaginationHistory({
               pagination_data: paginationDataset,