+++ src/services-hooks/useGetAllRequestLists.tsx
@@ -6,6 +6,7 @@ import {
   addToPaginationHistory,
 } from "../stores/apiData/requests-lists";
 import { pagination } from "../types/pagination";
+import ApiQueryParamsExtractor from "../utils/api-query-params-extractor";
 
 //axios instace interceptor for access token integration and refresh tokens
 export default function useGetAllRequestLists({
@@ -36,25 +37,28 @@ export default function useGetAllRequestLists({
   const [pagination, setPagination] = useState<pagination>({} as any);
   const getAllRequestList = useCallback(async () => {
     setIsLoading(true);
+    setIsFailed(false);
     try {
+      const { queryString, remakeRequest } = ApiQueryParamsExtractor({
+        dataset: {
+          page: search ? 1 : page,
+          start_date: start_date,
+          end_date: end_date,
+          sort: sort,
+          search: search,
+          limit,
+        },
+      });
       //check store if this requested data has been saved previously and retirve it
       //if not, make a new request and save into store
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (
-        foundPage &&
-        !(start_date && end_date) &&
-        !(sort === "asc") &&
-        !search &&
-        limit === 20
-      ) {
+      if (foundPage && !remakeRequest) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateRequestsLists({ data: foundPage?.data }));
       } else {
-        const response = await axios.get(
-          `admin/user-request?limit=${limit}&page=${page}`
-        );
+        const response = await axios.get(`admin/user-request?${queryString}`);
         const { user_requests } = response?.data?.data;
         const { data, current_page, last_page, per_page, total, from, to } =
           user_requests;
@@ -68,16 +72,16 @@ export default function useGetAllRequestLists({
           length: data?.length,
         };
         dispatch(updateRequestsLists({ data: data }));
-
-        dispatch(
-          addToPaginationHistory({
-            pagination_data: paginationDataset,
-            data,
-          })
-        );
+        if (!remakeRequest) {
+          dispatch(
+            addToPaginationHistory({
+              pagination_data: paginationDataset,
+              data,
+            })
+          );
+        }
         setPagination(paginationDataset);
       }
-      setIsLoading(false);
     } catch (error) {
       setIsFailed(true);
     } finally {