+++ src/services-hooks/useGetMaintenanceRequests.tsx
@@ -53,7 +53,7 @@ export default function useGetMaintenanceRequests({
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (foundPage && !remakeRequest && !(sort === "asc")) {
+      if (foundPage && !remakeRequest) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateRequisitionRequestList({ data: foundPage?.data }));
       } else {
@@ -73,7 +73,7 @@ export default function useGetMaintenanceRequests({
           length: data?.length,
         };
         dispatch(updateMaintenanceRequestList({ data }));
-        if (!search) {
+        if (!remakeRequest) {
           dispatch(
             addToPaginationHistory({
               pagination_data: paginationDataset,