+++ src/services-hooks/userGetReferral.tsx
@@ -53,7 +53,7 @@ export default function useGetReferrals({
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (foundPage && !remakeRequest && !(sort === "asc")) {
+      if (foundPage && !remakeRequest) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateReferralsList({ data: foundPage?.data }));
       } else {
@@ -65,7 +65,7 @@ export default function useGetReferrals({
         const data = sampleRReferralData?.data;
         const paginationDataset = sampleRReferralData?.pagination;
         dispatch(updateReferralsList({ data }));
-        if (!search) {
+        if (!remakeRequest) {
           dispatch(
             addToPaginationHistory({
               pagination_data: paginationDataset,