+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 17.03.2025
+
+> Calendar module Api integration
+> Api query params sort and limit logic update
+
 ##### 13.03.2025
 
 > Feature => Owner's report module UI development